add a brand new leveling system, loot hp and dmg now varies more precisely

+++ src/main/java/com/varyon/essence/PlayerEssenceData.java
@@ -2,5 +2,5 @@ package com.varyon.essence;
 
 import java.util.UUID;
 
-public record PlayerEssenceData(UUID uuid, String name, int essence) {
+public record PlayerEssenceData(UUID uuid, String name, double essence) {
 }
